Add test verifying dependency chains
Shared libraries require their (direct and indirect) dependencies to be built
in a position independent way. This might cause such a dependent library to be
built twice. As, however, shared libraries need not forward their static
dependencies this does not cause staging conflicts. Verify that this is also
the case for distant indirect dependencies.

Co-authored-by: Klaus Aehlig <klaus.aehlig@huawei.com>