rules CC: Make DEBUG variable a map
It expects now the key 'USE_DEBUG_FISSION', which enables debug
fission, but does not add any flags on its own. For this, the
'FISSION_CONFIG' key is expected with certain entries that provide
the compile and/or link flags that configure debug fission.

The flags are added only where needed, i.e., before running or
storing the respective compile or link actions.